Remove ANT+ and use BLE instead

+++ app/src/main/java/de/sualk1000/indoorcycler/BikeActivity.java
@@ -1,5 +1,8 @@
 package de.sualk1000.indoorcycler;
 
+
+import static android.os.Environment.DIRECTORY_DOWNLOADS;
+
 import android.content.Context;
 import android.os.AsyncTask;
 import android.util.Log;
@@ -21,6 +24,9 @@ import com.garmin.fit.RecordMesg;
 import com.garmin.fit.SessionMesg;
 import com.garmin.fit.Sport;
 import com.garmin.fit.SubSport;
+import com.garmin.fit.DeveloperDataIdMesg;
+import com.garmin.fit.DeveloperField;
+import com.garmin.fit.FieldDescriptionMesg;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -33,37 +39,40 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.math.BigDecimal;
+import java.nio.channels.FileChannel;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.LinkedList;
+import java.util.TimeZone;
 
 public class BikeActivity {
     private final static String TAG = BikeActivity.class.getSimpleName();
     private final IndoorCyclingService indoorCyclingService;
+    public double altitude;
 
     //Context context;
     //FragmentActivity activity;
     int lat = (int) (11930465 * 51.904491);
     int lon = (int) (11930465 * 10.427830);
-    Date start;
+    Date start = null;
+    boolean isEnded = false;
     //private EventMesg eventMesgStop;
     LinkedList<Mesg> messages = new LinkedList<Mesg>();
     Calendar cal = Calendar.getInstance();
     //EventMesg eventMesgStart = null;
     private double speed = -1;
-    private double power = -1;
+    double power = -1;
     private long distance = -1;
-    private int heartRate = 0;
+    int heartRate = 50;
 
     private long distance_offset = 0;
+
 
 
     BikeActivity(IndoorCyclingService indoorCyclingService)
     {
         this.indoorCyclingService = indoorCyclingService;
-        //this.context = context;
-        //this.activity = activity;
     }
 
     public void setSpeed(BigDecimal speed) {
@@ -82,7 +91,7 @@ public class BikeActivity {
     }
     SimpleDateFormat simpleDateFormat1 = new SimpleDateFormat("HH:mm:ss");
     SimpleDateFormat simpleDateFormat2 = new SimpleDateFormat("mm:ss");
-    public String getText() {
+    public String getText2() {
         String ret = "" ;
         if(start != null) {
             long time = (cal.getTime().getTime() - start.getTime());
@@ -99,9 +108,54 @@ public class BikeActivity {
         return ret;
     }
 
+    public String getHeartRateString() {
+        return String.format("%d hz", heartRate);
+    }
+    public String getPowerString() {
+        return String.format("%d W", new BigDecimal(power).intValue());
+    }
+    public String getSpeedString() {
+        int iSpeed = new BigDecimal(speed * 3.6).intValue();
+        return String.format("%d km/h",iSpeed);
+    }
+
+    public String getDistanceString() {
+        if(distance < 0)
+            return "-- m";
+        String ret = "" ;
+        if(distance + distance_offset < 1000)
+            ret += " " + (distance + distance_offset) + " m";
+        else
+        {
+            double kmDistance = (distance + distance_offset);
+            kmDistance = kmDistance / 1000;
+            ret += String.format("%3.2f",kmDistance) + " km";
+        }
+        return ret;
+    }
+
+    long getDuration()
+    {
+
+        if(start != null)
+            return (new Date().getTime() - start.getTime());
+        return 0;
+    }
+    public String getDurationString() {
+        String ret = "" ;
+        if(start != null) {
+            long time = (new Date().getTime() - start.getTime());
+            if(time > ( 3600 * 1000) )
+                ret += simpleDateFormat1.format(new Date(time));
+            else
+                ret += simpleDateFormat2.format(new Date(time));
+        }
+        return ret;
+    }
+
     private class sendDataTask extends AsyncTask<Void, Void, Void> {
 
-        private final File file;
+        private  File file;
 
         sendDataTask( File file) {
             this.file = file;
@@ -120,32 +174,52 @@ public class BikeActivity {
 
             try {
 
+                indoorCyclingService.sendShowWait("Login to Garmin");
                 GarminConnect gc = new GarminConnect();
-                if (gc.signin("sualk1000@googlemail.com", "4WQv9!q1") == false)
+                if (gc.signin("sualk1000@googlemail.com", "4WQv9!q1",UberManager.getInstance().getMainActivity()) == false)
                     throw new Exception("Login error");
 
 
 
 
 
+                if(messages.size() == 0)
+                {
+                    indoorCyclingService.sendTextMessage("Empty Message List");
+                    return null;
+                }
+                indoorCyclingService.sendShowWait("Upload Activity to Garmin");
 
-                File file = Export.buildFitFile((Context)indoorCyclingService, messages);
-                String upload_message = gc.uploadFitFile(file);
+                if(this.file == null || file.exists() == false) {
+                    this.file = Export.buildFitFile((Context) indoorCyclingService, messages);
+                }
+                Log.i(TAG, "Upload FIT file " + file.getAbsolutePath());
+                File downloadDirectory = indoorCyclingService.getExternalFilesDir(DIRECTORY_DOWNLOADS);
+                File outFile = new File(downloadDirectory.getAbsoluteFile() + "/test7.fit");
+
+
+                FileInputStream inStream = new FileInputStream(this.file);
+                FileOutputStream outStream = new FileOutputStream(outFile);
+                FileChannel inChannel = inStream.getChannel();
+                FileChannel outChannel = outStream.getChannel();
+                inChannel.transferTo(0, inChannel.size(), outChannel);
+                inStream.close();
+                outStream.close();
+                Log.i(TAG, "Copied FIT file to " + outFile.getAbsolutePath());
+                boolean upload_result = gc.uploadFitFile(file);
                 gc.close();
-                if (upload_message == null) {
-                    upload_message = "Upload ok";
-                    if(file != null & file.exists())
+                if (upload_result) {
+                    if(file != null && file.exists())
                         file.delete();
 
-                    indoorCyclingService.sendTextMessage(upload_message);
+                    indoorCyclingService.sendTextMessage("Upload OK");
 
                 }else
                 {
-                    indoorCyclingService.sendTextMessage(upload_message);
+                    indoorCyclingService.sendTextMessage("Error in Upload");
                     indoorCyclingService.sendCommand("ask");
 
                 }
-                Log.e(TAG, upload_message);
 
 
 
@@ -153,8 +227,16 @@ public class BikeActivity {
                 Log.e(TAG, e.toString());
 
                 indoorCyclingService.sendTextMessage("Failed: " + e.getMessage());
-                return null;
+
+
             }
+            isEnded = false;
+            distance = 0;
+            distance_offset = 0;
+
+            indoorCyclingService.sendHideWait();
+            indoorCyclingService.sendControlStatus("start",true);
+            indoorCyclingService.sendControlStatus("stop",false);
 
             return null;
         }
@@ -164,52 +246,34 @@ public class BikeActivity {
     void addStartMessage(Date startDate)
     {
 
+
         cal.setTime(startDate);
-        FileIdMesg fileIdMesg = new FileIdMesg();
-        fileIdMesg.setType(com.garmin.fit.File.ACTIVITY);
-        fileIdMesg.setManufacturer(Manufacturer.GARMIN);
-        fileIdMesg.setProduct(1);
-        fileIdMesg.setSerialNumber(3986196270L);
-        fileIdMesg.setTimeCreated(new DateTime(startDate));
-
-        messages.add(fileIdMesg);
-
-
-        FileCreatorMesg creatorMesg = new FileCreatorMesg();
-        creatorMesg.setSoftwareVersion(700);
-
-
-        DeviceInfoMesg deviceInfoMesg = new DeviceInfoMesg();
-        deviceInfoMesg.setDeviceIndex((short) 1);
-        deviceInfoMesg.setManufacturer(Manufacturer.GARMIN);
-        deviceInfoMesg.setProduct(2888);
-        deviceInfoMesg.setProductName("FIT Cookbook"); // Max 20 Chars
-        deviceInfoMesg.setSerialNumber(1L);
-        deviceInfoMesg.setSoftwareVersion((float) 7L);
-        deviceInfoMesg.setTimestamp(new DateTime(startDate));
-
-        deviceInfoMesg.setDeviceIndex(DeviceIndex.CREATOR);
-        deviceInfoMesg.setManufacturer(Manufacturer.DEVELOPMENT);
-
-
-        messages.add(deviceInfoMesg);
 
         EventMesg eventMesgStart = new EventMesg();
         eventMesgStart.setTimestamp(new DateTime(cal.getTime()));
         eventMesgStart.setEvent(Event.TIMER);
         eventMesgStart.setEventType(EventType.START);
+
+        Log.i(TAG, "Add Message " + eventMesgStart.getClass().getSimpleName() + eventMesgStart.getName());
+
         messages.add(eventMesgStart);
+
+
 
     }
     public boolean isStarted()
     {
         if(start != null)
             return true;
+        /*
         File file = new File(indoorCyclingService.getCacheDir() , "MyCache.json");
         if(file.exists() == false)
             return false;
 
         return resume(file);
+
+         */
+        return false;
     }
 
     public boolean resume(File file) {
@@ -223,7 +287,7 @@ public class BikeActivity {
 
             Date startDate = null;
             Date endDate = null;
-            long distance = 0;
+            long distance2 = 0;
             while (lineData != null) {
 
                 Log.i(TAG, lineData);
@@ -236,9 +300,9 @@ public class BikeActivity {
                     startDate = date;
                 endDate = date;
                 recordMesg.setTimestamp(new DateTime(date));
-                distance =  Long.valueOf(jsonObject.get("distance").toString());
+                distance2 =  Long.valueOf(jsonObject.get("distance").toString());
 
-                recordMesg.setDistance((float) distance); // Ramp
+                recordMesg.setDistance((float) distance2); // Ramp
                 recordMesg.setSpeed(Float.valueOf(jsonObject.get("speed").toString())); // Speed in m/s = km/h / 3.6
                 recordMesg.setPower(Float.valueOf(jsonObject.get("power").toString()).intValue()); // Watt
                 if(jsonObject.has("heartRate")) {
@@ -250,7 +314,8 @@ public class BikeActivity {
                 recordMesg.setPositionLat(lat);
                 recordMesg.setPositionLong(lon);
                 addMessage(recordMesg,date);
-                messages.add(recordMesg);
+
+                //messages.add(recordMesg);
 
                 //retBuf.append(lineData);
                 lineData = bufferedReader.readLine();
@@ -275,16 +340,30 @@ public class BikeActivity {
     }
 
     public void start() {
+
+        if(isEnded)
+        {
+            indoorCyclingService.sendTextMessage("Activity is ended.");
+            return;
+
+        }
         if(start != null)
         {
             indoorCyclingService.sendTextMessage("Already started");
 
-            return;
+        }else
+        {
+            start = new Date();
+            cal.setTime(start);
+
+            if(messages.size() > 0)
+            {
+                messages.clear();
+            }
+            addStartMessage(start);
+
         }
 
-        start = new Date();
-        cal.setTime(start);
-        addStartMessage(start);
 
         indoorCyclingService.sendControlStatus("start",false);
         indoorCyclingService.sendControlStatus("stop",true);
@@ -294,22 +373,32 @@ public class BikeActivity {
 
     void addEndMessage(Date startDate,Date endDate,int lat,int lon,double distance)
     {
+        if(isEnded)
+        {
+            indoorCyclingService.sendTextMessage("Activity is ended.");
+            return;
+
+        }
+
+        isEnded = true;
         EventMesg eventMesgStop = new EventMesg();
         cal.setTime(endDate);
 
-        eventMesgStop.setTimestamp(new DateTime(cal.getTime()));
+        eventMesgStop.setTimestamp(new DateTime(endDate));
         eventMesgStop.setEvent(Event.TIMER);
         eventMesgStop.setEventType(EventType.STOP_ALL);
+        Log.i(TAG, "addEndMessage Add Message " + eventMesgStop.getClass().getSimpleName() + eventMesgStop.getName());
+
         messages.add(eventMesgStop);
 
         LapMesg lapMesg = new LapMesg();
         lapMesg.setMessageIndex(0);
-        lapMesg.setStartTime(new DateTime(cal.getTime()));
-        lapMesg.setTimestamp(new DateTime(cal.getTime()));
+        lapMesg.setStartTime(new DateTime(new DateTime(startDate)));
+        lapMesg.setTimestamp(new DateTime(new DateTime(endDate)));
 
         long time = (cal.getTime().getTime() - startDate.getTime()) / 1000;
-        lapMesg.setTotalElapsedTime((float) (new DateTime(cal.getTime()).getTimestamp() - new DateTime(startDate).getTimestamp()));
-        lapMesg.setTotalTimerTime((float) (new DateTime(cal.getTime()).getTimestamp() - new DateTime(startDate).getTimestamp()));
+        lapMesg.setTotalElapsedTime((float) (new DateTime(endDate).getTimestamp() - new DateTime(startDate).getTimestamp()));
+        lapMesg.setTotalTimerTime((float) (new DateTime(endDate).getTimestamp() - new DateTime(startDate).getTimestamp()));
 
         lapMesg.setStartPositionLat((int) lat);
         lapMesg.setStartPositionLong((int) lon);
@@ -318,6 +407,7 @@ public class BikeActivity {
         lapMesg.setTotalDistance((float) distance);
 
         lapMesg.setAvgSpeed((float) (distance / time));
+        Log.i(TAG, "addEndMessage Add Message " + lapMesg.getClass().getSimpleName() + lapMesg.getName());
         messages.add(lapMesg);
 
         SessionMesg sessionMesg = new SessionMesg();
@@ -326,31 +416,35 @@ public class BikeActivity {
         sessionMesg.setStartTime(new DateTime(startDate));
         sessionMesg.setEvent(Event.LAP);
         sessionMesg.setEventType(EventType.STOP);
-        sessionMesg.setTotalElapsedTime((float) (new DateTime(cal.getTime()).getTimestamp() - new DateTime(startDate).getTimestamp()));
-        sessionMesg.setTotalTimerTime((float) (new DateTime(cal.getTime()).getTimestamp() - new DateTime(startDate).getTimestamp()));
+        sessionMesg.setTotalElapsedTime((float) (new DateTime(endDate).getTimestamp() - new DateTime(startDate).getTimestamp()));
+        sessionMesg.setTotalTimerTime((float) (new DateTime(endDate).getTimestamp() - new DateTime(startDate).getTimestamp()));
         sessionMesg.setSport(Sport.CYCLING);
         sessionMesg.setSubSport(SubSport.INDOOR_CYCLING);
         sessionMesg.setFirstLapIndex(0);
         sessionMesg.setNumLaps(1);
         sessionMesg.setStartPositionLat(lat);
         sessionMesg.setStartPositionLong(lon);
+        Log.i(TAG, "addEndMessage Add Message " + sessionMesg.getClass().getSimpleName() + sessionMesg.getName());
         messages.add(sessionMesg);
 
 
         ActivityMesg activityMesg = new ActivityMesg();
         activityMesg.setTimestamp(new DateTime(cal.getTime()));
         activityMesg.setNumSessions(1);
-        activityMesg.setLocalTimestamp((new DateTime(endDate).getTimestamp()));
-        activityMesg.setTotalTimerTime(sessionMesg.getTotalTimerTime());
-        activityMesg.setEvent(Event.ACTIVITY);
-        activityMesg.setEventType(EventType.STOP);
-        activityMesg.setType(Activity.MANUAL);
+
+        TimeZone timeZone = TimeZone.getDefault();
+        long timezoneOffset = (timeZone.getRawOffset() + timeZone.getDSTSavings()) / 1000;
+        activityMesg.setLocalTimestamp(new DateTime(endDate).getTimestamp() + timezoneOffset);
+        activityMesg.setTotalTimerTime((float) (new DateTime(endDate).getTimestamp() - new DateTime(startDate).getTimestamp()));
+
+        //activityMesg.setEvent(Event.ACTIVITY);
+        //activityMesg.setEventType(EventType.STOP);
+        //activityMesg.setType(Activity.MANUAL);
 
 
+        Log.i(TAG, "addEndMessage Add Message " + activityMesg.getClass().getSimpleName() + activityMesg.getName());
         messages.add(activityMesg);
 
-        start = null;
-        distance_offset = 0;
     }
     public void stop() {
         if(start == null)
@@ -361,8 +455,31 @@ public class BikeActivity {
             return;
         }
 
+        indoorCyclingService.sendControlStatus("start",false);
+        indoorCyclingService.sendControlStatus("stop",false);
+
+
         addEndMessage(start,new Date(),lat,lon,distance+distance_offset);
 
+        if(getDuration() < 60*1000)
+        {
+            indoorCyclingService.sendTextMessage("Activity to short.");
+            isEnded = false;
+            start = null;
+            distance = 0;
+            distance_offset = 0;
+
+            File file = new File(indoorCyclingService.getCacheDir() , "MyCache.json");
+            if(file.exists())
+                file.delete();
+
+            indoorCyclingService.sendHideWait();
+            indoorCyclingService.sendControlStatus("start",true);
+            indoorCyclingService.sendControlStatus("stop",false);
+
+            return;
+        }
+        start = null;
         send(null);
 
 
@@ -371,7 +488,7 @@ public class BikeActivity {
 
     }
     public void send(File file) {
-        if(start != null && file == null)
+        if(isEnded == false)
         {
             indoorCyclingService.sendTextMessage("Not stopped");
 
@@ -382,9 +499,17 @@ public class BikeActivity {
     }
     public void addMessage(RecordMesg recordMesg,Date date) {
 
+        if(isEnded)
+        {
+            indoorCyclingService.sendTextMessage("Activity is ended.");
+            return;
+
+        }
+
         if(messages.size() == 0) {
             addStartMessage(date);
         }
+        Log.i(TAG, "addMessage(...) Add Message " + recordMesg.getClass().getSimpleName() + recordMesg.getName());
         messages.add(recordMesg);
     }
 
@@ -425,7 +550,7 @@ public class BikeActivity {
                 recordMesg.setPositionLat(lat);
                 recordMesg.setPositionLong(lon);
                 addMessage(recordMesg,date);
-                messages.add(recordMesg);
+                //messages.add(recordMesg);
 
                 //retBuf.append(lineData);
                 lineData = bufferedReader.readLine();
@@ -443,7 +568,7 @@ public class BikeActivity {
 
     public void addMessage() {
 
-        if(start == null)
+        if(start == null || isEnded)
         {
             return;
         }
@@ -498,6 +623,9 @@ public class BikeActivity {
             //recordMesg.setAltitude(200F); // Triangle
             recordMesg.setPositionLat(lat);
             recordMesg.setPositionLong(lon);
+
+        Log.i(TAG, "addMessage Add Message " + recordMesg.getClass().getSimpleName() + recordMesg.getName());
+
         messages.add(recordMesg);
 
     }
